workers/requirements.txt

celery==5.5.3
elasticsearch==7.13.4
fastapi==0.116.1
httpx==0.28.1
passlib==1.7.4
pika==1.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.2.0
requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.35.0


opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0


opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-celery==0.45b0
